Introduce SubmitButton, and move buttons to separate dir
This will be used for assessment.
