+++ src/buttons/BackButton.js
@@ -11,7 +11,7 @@ export default class BackButton extends React.Component {
                     onClick={this.onClick.bind(this)}>
                 <svg className="octicon octicon-arrow-left octicon-before" viewBox="0 0 10 16" version="1.1" width="10" height="16" aria-hidden="true"><path fillRule="evenodd" d="M6 3L0 8l6 5v-3h4V6H6z"></path></svg>
                 Back</button>
-        )
+        );
     }
     onClick() {
         this.props.onClick();
@@ -21,4 +21,4 @@ export default class BackButton extends React.Component {
 BackButton.propTypes = {
     onClick: PropTypes.func.isRequired,
     showing: PropTypes.bool.isRequired
-}
+};
